Add unit test for cmd stdout assertions
