+++ unit/testing-tf/main.tf
@@ -11,6 +11,19 @@ terraform {
 provider "random" {
 }
 
+variable "length" {
+  type    = number
+  default = 16
+  validation {
+    condition     = var.length == 16
+    error_message = "Random string length must be 16."
+  }
+}
+
 resource "random_string" "s" {
-  length = 16
+  length = var.length
+}
+
+output "random_string_result" {
+  value = random_string.s.result
 }
