+++ unit/unit_test.go
@@ -1,6 +1,7 @@
 package unit
 
 import (
+	"strings"
 	"testing"
 
 	"github.com/gruntwork-io/terratest/modules/terraform"
@@ -8,11 +9,10 @@ import (
 
 // the workspace cannot be deleted in the case that the "before" workspace is the same
 // as the one used for the test. If this test passes, then the behavior is correct
-func TestSupportForUnitTestWitNoWorkspaceChange(t *testing.T) {
+func TestSupportForUnitTestWithNoWorkspaceChange(t *testing.T) {
 	tfOptions := &terraform.Options{
 		TerraformDir: "testing-tf/",
 		Upgrade:      true,
-		Vars:         map[string]interface{}{},
 	}
 
 	testFixture := UnitTestFixture{
@@ -21,6 +21,33 @@ func TestSupportForUnitTestWitNoWorkspaceChange(t *testing.T) {
 		ExpectedResourceCount:           1,
 		ExpectedResourceAttributeValues: nil,
 		Workspace:                       "default",
+	}
+
+	RunUnitTests(&testFixture)
+}
+
+func TestTerraformCommandStdout(t *testing.T) {
+	tfOptions := &terraform.Options{
+		TerraformDir: "testing-tf/",
+		Upgrade:      true,
+		Vars: map[string]interface{}{
+			"length": 15,
+		},
+	}
+
+	testFixture := UnitTestFixture{
+		GoTest:    t,
+		TfOptions: tfOptions,
+		CommandStdoutAssertions: []TerraformCommandStdoutValidation{
+			func(t *testing.T, output string, err error) {
+				if err == nil {
+					t.Fatal("should err")
+				}
+				if !strings.Contains(output, "Invalid value for variable") {
+					t.Fatal("should contain 'Invalid value for variable'")
+				}
+			},
+		},
 	}
 
 	RunUnitTests(&testFixture)
